+++ packages/eslint-config-vue/vue2-typescript.js
@@ -1,5 +1,7 @@
 module.exports = {
-  extends: ['./rules/vue2', '@sxwy/eslint-config-base/typescript'].map(
-    require.resolve
-  )
+  extends: [
+    './rules/vue2-base',
+    '@sxwy/eslint-config-base/typescript',
+    './rules/vue2-parser'
+  ].map(require.resolve)
 }