+++ packages/eslint-config-vue/vue2.js
@@ -1,3 +1,7 @@
 module.exports = {
-  extends: ['./rules/vue2', '@sxwy/eslint-config-base'].map(require.resolve)
+  extends: [
+    './rules/vue2-base',
+    '@sxwy/eslint-config-base',
+    './rules/vue2-parser'
+  ].map(require.resolve)
 }